P1-ZAD2 Added categories with id and name, all categories can be displayed.
