+++ src/main/java/pl/akalkowski/demo/controllers/GifController.java
@@ -1,10 +1,10 @@
-package pl.akalkowski.demo.controller;
+package pl.akalkowski.demo.controllers;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.ModelMap;
 import org.springframework.web.bind.annotation.GetMapping;
-import pl.akalkowski.demo.repository.GifRepository;
+import pl.akalkowski.demo.repositories.GifRepository;
 
 /**
  * Created by monik on 04.11.2017.
